minor+refactor: Moved DockerTestContainerManager
